requirements.txt
PySimpleGUI>=4.60.5
pandas>=2.0.1
tkcalendar>=1.6.1
openpyxl>=3.1.2
xlwings>=0.30.7
deepdiff>=8.0.1
mysql-connector-python>=9.1.0